Organized e2e environment managers into service-managers folder (#25291)
no_ref
- Moved managers (Ghost, MySQL, Portal, Tinybird) into service-managers/subfolder for better code organization
- It aligns better with Docker Compose terminology, and separates cleaner environment manager, compose from service managers
