+++ e2e/helpers/environment/EnvironmentManager.ts
@@ -3,22 +3,10 @@ import baseDebug from '@tryghost/debug';
 import logging from '@tryghost/logging';
 import {DOCKER_COMPOSE_CONFIG, PORTAL, TINYBIRD} from './constants';
 import {DockerCompose} from './DockerCompose';
-import {GhostManager} from './GhostManager';
-import {MySQLManager} from './MySQLManager';
-import {PortalManager} from './PortalManager';
-import {TinybirdManager} from './TinybirdManager';
+import {GhostInstance, GhostManager, MySQLManager, PortalManager, TinybirdManager} from './service-managers';
 import {randomUUID} from 'crypto';
 
 const debug = baseDebug('e2e:EnvironmentManager');
-
-export interface GhostInstance {
-    containerId: string; // docker container ID
-    instanceId: string; // unique instance name (e.g. ghost_<siteUuid>)
-    database: string;
-    port: number;
-    baseUrl: string;
-    siteUuid: string;
-}
 
 /**
  * Manages the lifecycle of Docker containers and shared services for end-to-end tests
@@ -73,21 +61,7 @@ export class EnvironmentManager {
         logging.info('Starting global environment setup...');
 
         await this.cleanupResources();
-
-        // Start required docker compose services
-        this.dockerCompose.up([
-            'mysql',
-            'ghost-migrations',
-            'caddy',
-            'analytics',
-            'tinybird-local',
-            'tb-cli',
-            'mailpit',
-            'portal'
-        ]);
-
-        await this.dockerCompose.waitForAll();
-
+        await this.dockerCompose.up();
         await this.mysql.createSnapshot();
         this.tinybird.fetchAndSaveConfig();
