+++ e2e/helpers/environment/index.ts
@@ -1,1 +1,3 @@
+export * from './service-managers';
 export * from './EnvironmentManager';
+
